+++ src/gui/session/SessionSideBar.h
@@ -11,7 +11,6 @@
 class QLabel;
 class QTabWidget;
 class Session;
-class ShellListWidget;
 
 class SessionSideBar final : public QWidget
 {
@@ -23,16 +22,16 @@ public:
     void bindSession(Session *session);
     void unbindSession();
 
-    ShellListWidget *shellList() const { return m_shellList; }
     QWidget *fileContainer() const { return m_fileContainer; }
     QWidget *tunnelContainer() const { return m_tunnelContainer; }
 
 private:
     void loadTabState();
     void saveTabState();
+    QString currentTabId() const;
+    void setCurrentTabId(const QString &tabId);
 
     QTabWidget *m_tabs = nullptr;
-    ShellListWidget *m_shellList = nullptr;
     QWidget *m_fileContainer = nullptr;
     QWidget *m_tunnelContainer = nullptr;
     QLabel *m_placeholder = nullptr;